+++ math/math.go
@@ -2,6 +2,6 @@ package math
 
 import "math"
 
-func Clamp(value float64, min float64, max float64) float64 {
+func Clamp(value float32, min float32, max float32) float32 {
 	return math.Max(min, math.Min(max, value))
 }
