+++ math/sphere.go
@@ -2,5 +2,5 @@ package math
 
 type Sphere struct {
 	center *Vector3
-	radius float64
+	radius float32
 }
