+++ math/spherical.go
@@ -3,16 +3,16 @@ package math
 import "math"
 
 type Spherical struct {
-	radius float64
-	phi    float64 // polar angle
-	theta  float64 // azimuthal angle
+	radius float32
+	phi    float32 // polar angle
+	theta  float32 // azimuthal angle
 }
 
 func NewDefaultSpherical() *Spherical {
 	return NewSpherical(1, 0, 0)
 }
 
-func NewSpherical(radius float64, phi float64, theta float64) *Spherical {
+func NewSpherical(radius float32, phi float32, theta float32) *Spherical {
 	return &Spherical{
 		radius: radius,
 		phi:    phi,
@@ -20,7 +20,7 @@ func NewSpherical(radius float64, phi float64, theta float64) *Spherical {
 	}
 }
 
-func (sphere *Spherical) Set(radius float64, phi float64, theta float64) {
+func (sphere *Spherical) Set(radius float32, phi float32, theta float32) {
 	sphere.radius = radius
 	sphere.phi = phi
 	sphere.theta = theta
@@ -41,7 +41,7 @@ func (sphere *Spherical) Copy(other *Spherical) {
 }
 
 func (sphere *Spherical) MakeSafe() {
-	const EPS = float64(0.000001)
+	const EPS = float32(0.000001)
 	sphere.phi = math.Max(EPS, math.Min(math.Pi-EPS, sphere.phi))
 }
 
@@ -49,7 +49,7 @@ func (sphere *Spherical) SetFromVector3(vec *Vector3) {
 	sphere.SetFromCartesianCoordinates(vec.X, vec.Y, vec.Z)
 }
 
-func (sphere *Spherical) SetFromCartesianCoordinates(x float64, y float64, z float64) {
+func (sphere *Spherical) SetFromCartesianCoordinates(x float32, y float32, z float32) {
 	sphere.radius = math.Sqrt(x*x + y*y + z*z)
 
 	if sphere.radius == 0 {
